bots: Stop provisioning bot dependencies.
So far, there are only two bots with dependencies (define and
thesaurus). Since there are not many bots that require extra
bot-specific dependencies, we now specify these in setup.py so
that they are installed as part of the package. This move
obsoltes bots_api/provision.py and the --provision option for
bots_api/run.py.