bots_api: Expose run.py as an executable named zulip-run-bot.
We now expose run.py as an executable named zulip-run-bot that
will be installed as a part of our bots package. Now, we can
also just pass the name of the bot to zulip-run-bot to run it,
instead of having to pass a path to the source file containing
the handler_class.

An additional benefit of this new approach is that we can stop
dealing with file paths and validating them and sanitizing them
and what not. It makes up for a much cleaner and simpler
implementation by removing a lot of manual modifications of
sys.path in order to import bots.